Create a useOperation
